+++ src/roq/okx/gateway/market_data.cpp
@@ -87,6 +87,7 @@ MarketData::MarketData(Handler &handler, io::Context &context, uint16_t stream_i
           .subscribe = create_metrics(shared.settings, name_, "subscribe"sv),
           .unsubscribe = create_metrics(shared.settings, name_, "unsubscribe"sv),
           .notice = create_metrics(shared.settings, name_, "notice"sv),
+          .status = create_metrics(shared.settings, name_, "status"sv),
           .login = create_metrics(shared.settings, name_, "login"sv),
           .estimated_price = create_metrics(shared.settings, name_, "estimated_price"sv),
           .price_limit = create_metrics(shared.settings, name_, "price_limit"sv),
@@ -131,6 +132,7 @@ void MarketData::operator()(metrics::Writer &writer) const {
       .write(profile_.subscribe, metrics::Type::PROFILE)
       .write(profile_.unsubscribe, metrics::Type::PROFILE)
       .write(profile_.notice, metrics::Type::PROFILE)
+      .write(profile_.status, metrics::Type::PROFILE)
       .write(profile_.login, metrics::Type::PROFILE)
       .write(profile_.estimated_price, metrics::Type::PROFILE)
       .write(profile_.price_limit, metrics::Type::PROFILE)
@@ -387,8 +389,18 @@ void MarketData::operator()(Trace<protocol::json::Notice> const &event) {
   });
 }
 
-void MarketData::operator()(Trace<protocol::json::Status> const &) {
-  log::fatal("Unexpected"sv);
+void MarketData::operator()(Trace<protocol::json::Status> const &event) {
+  profile_.status([&]() {
+    auto &[trace_info, status] = event;
+    log::warn("status={}"sv, status);
+    for (auto &item : status.data) {
+      if (item.service_type == 0 && item.state == protocol::json::State::ONGOING) {
+        log::warn("*** DISCONNECT: ONGOING MAINTENANCE ***"sv);
+        (*connection_).close();
+        return;
+      }
+    }
+  });
 }
 
 void MarketData::operator()(Trace<protocol::json::Instruments> const &) {